+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is a changelog for **ZachranObed** application.
 - **ZOB-422** Add optional pickup confirmation.
 - **ZOB-421** Add support for dynamic banners.
 - **ZOB-424** Add optional manual delivery flow.
+- **ZOB-436** Add box return confirmation for canteens.
 
 ### Changed
 - **ZOB-438** Add pacing and retry logic for monthly reports, fix filename and add support for custom time ranges.